[IMP] web: string domain function handler
When getting a domain from activeFields in Owl, sometimes the value passed is not a list or function.

When the field mentioned have a `check_company=True` in the initialization, the domain passed will be in a form of string instead of a list.

This commit aims to publicize the domain handler in Field so that it can be used anywhere by calling the function, as getting field's domain is getting increasingly common

Task-id: 3549961
Part-of: odoo/odoo#140773